+++ sdk/cosmosdb/cosmos/src/queryExecutionContext/documentProducer.ts
@@ -13,11 +13,11 @@ import {
 import type { DiagnosticNodeInternal } from "../diagnostics/DiagnosticNodeInternal";
 import type { FeedOptions } from "../request";
 import type { Response } from "../request";
-import type { FetchFunctionCallback } from "./defaultQueryExecutionContext";
 import { DefaultQueryExecutionContext } from "./defaultQueryExecutionContext";
+import type { FetchFunctionCallback } from "./defaultQueryExecutionContext";
 import { FetchResult, FetchResultType } from "./FetchResult";
-import type { CosmosHeaders } from "./headerUtils";
 import { getInitialHeader, mergeHeaders } from "./headerUtils";
+import type { CosmosHeaders } from "./headerUtils";
 import type { SqlQuerySpec } from "./index";
 
 /** @hidden */
@@ -33,6 +33,9 @@ export class DocumentProducer {
   public generation: number = 0;
   private respHeaders: CosmosHeaders;
   private internalExecutionContext: DefaultQueryExecutionContext;
+  public startEpk: string;
+  public endEpk: string;
+  public populateEpkRangeHeaders: boolean;
 
   /**
    * Provides the Target Partition Range Query Execution Context.
@@ -49,6 +52,9 @@ export class DocumentProducer {
     targetPartitionKeyRange: PartitionKeyRange,
     options: FeedOptions,
     correlatedActivityId: string,
+    startEpk?: string,
+    endEpk?: string,
+    populateEpkRangeHeaders?: boolean,
   ) {
     // TODO: any options
     this.collectionLink = collectionLink;
@@ -68,12 +74,10 @@ export class DocumentProducer {
       this.fetchFunction,
       correlatedActivityId,
     );
+    this.startEpk = startEpk;
+    this.endEpk = endEpk;
+    this.populateEpkRangeHeaders = populateEpkRangeHeaders;
   }
-  /**
-   * Synchronously gives the contiguous buffered results (stops at the first non result) if any
-   * @returns buffered current items if any
-   * @hidden
-   */
   public peekBufferedItems(): any[] {
     const bufferedResults = [];
     for (let i = 0, done = false; i < this.fetchResults.length && !done; i++) {
@@ -101,6 +105,8 @@ export class DocumentProducer {
     const path = getPathFromLink(this.collectionLink, ResourceType.item);
     diagnosticNode.addData({ partitionKeyRangeId: this.targetPartitionKeyRange.id });
     const id = getIdFromLink(this.collectionLink);
+    const startEpk = this.populateEpkRangeHeaders ? this.startEpk : undefined;
+    const endEpk = this.populateEpkRangeHeaders ? this.endEpk : undefined;
 
     return this.clientContext.queryFeed({
       path,
@@ -112,6 +118,8 @@ export class DocumentProducer {
       diagnosticNode,
       partitionKeyRangeId: this.targetPartitionKeyRange["id"],
       correlatedActivityId: correlatedActivityId,
+      startEpk: startEpk,
+      endEpk: endEpk,
     });
   };
 
@@ -120,13 +128,14 @@ export class DocumentProducer {
   }
 
   public gotSplit(): boolean {
-    const fetchResult = this.fetchResults[0];
-    if (fetchResult.fetchResultType === FetchResultType.Exception) {
-      if (DocumentProducer._needPartitionKeyRangeCacheRefresh(fetchResult.error)) {
-        return true;
+    if (this.fetchResults.length !== 0) {
+      const fetchResult = this.fetchResults[0];
+      if (fetchResult.fetchResultType === FetchResultType.Exception) {
+        if (DocumentProducer._needPartitionKeyRangeCacheRefresh(fetchResult.error)) {
+          return true;
+        }
       }
     }
-
     return false;
   }
 
@@ -137,7 +146,6 @@ export class DocumentProducer {
   }
 
   private _updateStates(err: any, allFetched: boolean): void {
-    // TODO: any Error
     if (err) {
       this.err = err;
       return;
@@ -163,9 +171,9 @@ export class DocumentProducer {
   }
 
   /**
-   * Fetches and bufferes the next page of results and executes the given callback
+   * Fetches and bufferes the next page of results in internal buffer
    */
-  public async bufferMore(diagnosticNode: DiagnosticNodeInternal): Promise<Response<any>> {
+  public async bufferMore(diagnosticNode: DiagnosticNodeInternal): Promise<void> {
     if (this.err) {
       throw this.err;
     }
@@ -176,10 +184,17 @@ export class DocumentProducer {
       ++this.generation;
       this._updateStates(undefined, resources === undefined);
       if (resources !== undefined) {
-        // some more results
+        // add fetched header to the 1st element in the buffer
+        let addHeaderToFetchResult = true;
         resources.forEach((element: any) => {
-          // TODO: resources any
-          this.fetchResults.push(new FetchResult(element, undefined));
+          this.fetchResults.push(
+            new FetchResult(
+              element,
+              undefined,
+              addHeaderToFetchResult ? headerResponse : getInitialHeader(),
+            ),
+          );
+          addHeaderToFetchResult = false;
         });
       }
 
@@ -193,20 +208,14 @@ export class DocumentProducer {
         headerResponse[Constants.HttpHeaders.QueryMetrics][this.targetPartitionKeyRange.id] =
           queryMetrics;
       }
-
-      return { result: resources, headers: headerResponse };
+      mergeHeaders(this.respHeaders, headerResponse);
     } catch (err: any) {
-      // TODO: any error
       if (DocumentProducer._needPartitionKeyRangeCacheRefresh(err)) {
         // Split just happend
         // Buffer the error so the execution context can still get the feedResponses in the itemBuffer
         const bufferedError = new FetchResult(undefined, err);
         this.fetchResults.push(bufferedError);
-        // Putting a dummy result so that the rest of code flows
-        return {
-          result: [bufferedError],
-          headers: err.headers,
-        };
+        mergeHeaders(this.respHeaders, err.headers);
       } else {
         this._updateStates(err, err.resources === undefined);
         throw err;
@@ -214,41 +223,81 @@ export class DocumentProducer {
     }
   }
 
-  /**
-   * Synchronously gives the bufferend current item if any
-   * @returns buffered current item if any
-   * @hidden
-   */
   public getTargetParitionKeyRange(): PartitionKeyRange {
     return this.targetPartitionKeyRange;
   }
+  /**
+   * Peak the next item in the buffer
+   */
+  public peakNextItem(): any {
+    if (this.err) {
+      throw this.err;
+    }
+    if (this.allFetched || this.fetchResults.length === 0) {
+      return undefined;
+    }
+    const fetchResult = this.fetchResults[0];
+
+    switch (fetchResult.fetchResultType) {
+      case FetchResultType.Done:
+        return undefined;
+
+      case FetchResultType.Exception: // do not throw this error
+        return undefined;
+
+      case FetchResultType.Result:
+        return fetchResult.feedResponse;
+    }
+  }
 
   /**
-   * Fetches the next element in the DocumentProducer.
+   * Returns the first item in the buffered results if any, or [] otherwise.
    */
-  public async nextItem(diagnosticNode: DiagnosticNodeInternal): Promise<Response<any>> {
+  public async fetchNextItem(): Promise<Response<any>> {
     if (this.err) {
       this._updateStates(this.err, undefined);
       throw this.err;
     }
-
+    if (this.allFetched) {
+      return { result: undefined, headers: this._getAndResetActiveResponseHeaders() };
+    }
     try {
-      const { result, headers } = await this.current(diagnosticNode);
-
-      const fetchResult = this.fetchResults.shift();
+      const { result, headers } = await this.current();
       this._updateStates(undefined, result === undefined);
-      if (fetchResult.feedResponse !== result) {
-        throw new Error(`Expected ${fetchResult.feedResponse} to equal ${result}`);
+      if (result === undefined || result.length === 0) {
+        return { result: undefined, headers };
       }
-      switch (fetchResult.fetchResultType) {
-        case FetchResultType.Done:
-          return { result: undefined, headers };
-        case FetchResultType.Exception:
-          fetchResult.error.headers = headers;
-          throw fetchResult.error;
-        case FetchResultType.Result:
-          return { result: fetchResult.feedResponse, headers };
+      return { result, headers }; //
+    } catch (err: any) {
+      this._updateStates(err, err.item === undefined);
+      throw err;
+    }
+  }
+  /**
+   * Fetches all the buffered results
+   */
+  public async fetchBufferedItems(): Promise<Response<any[]>> {
+    if (this.err) {
+      this._updateStates(this.err, undefined);
+      throw this.err;
+    }
+    if (this.allFetched) {
+      return { result: undefined, headers: this._getAndResetActiveResponseHeaders() };
+    }
+    const resources: any[] = [];
+    const resHeaders: CosmosHeaders = getInitialHeader();
+    try {
+      while (this.fetchResults.length > 0) {
+        const { result, headers } = await this.current();
+        this._updateStates(undefined, result === undefined);
+        mergeHeaders(resHeaders, headers);
+        if (result === undefined) {
+          return { result: resources.length > 0 ? resources : undefined, headers: resHeaders };
+        } else {
+          resources.push(result);
+        }
       }
+      return { result: resources, headers: resHeaders };
     } catch (err: any) {
       this._updateStates(err, err.item === undefined);
       throw err;
@@ -258,10 +307,10 @@ export class DocumentProducer {
   /**
    * Retrieve the current element on the DocumentProducer.
    */
-  public async current(diagnosticNode: DiagnosticNodeInternal): Promise<Response<any>> {
+  private async current(): Promise<Response<any>> {
     // If something is buffered just give that
     if (this.fetchResults.length > 0) {
-      const fetchResult = this.fetchResults[0];
+      const fetchResult = this.fetchResults.shift();
       // Need to unwrap fetch results
       switch (fetchResult.fetchResultType) {
         case FetchResultType.Done:
@@ -288,12 +337,7 @@ export class DocumentProducer {
       };
     }
 
-    // If there are no more bufferd items and there are still items to be fetched then buffer more
-    const { result, headers } = await this.bufferMore(diagnosticNode);
-    mergeHeaders(this.respHeaders, headers);
-    if (result === undefined) {
-      return { result: undefined, headers: this.respHeaders };
-    }
-    return this.current(diagnosticNode);
+    // If the internal buffer is empty, return empty result
+    return { result: [], headers: this._getAndResetActiveResponseHeaders() };
   }
 }
