+++ sdk/cosmosdb/cosmos/src/queryExecutionContext/index.ts
@@ -4,7 +4,7 @@ export * from "./headerUtils";
 export * from "./SqlQuerySpec";
 export * from "./defaultQueryExecutionContext";
 export * from "./Aggregators";
-// export * from "./documentProducer";
+export * from "./documentProducer";
 export * from "./FetchResult";
 export * from "./orderByDocumentProducerComparator";
 export * from "./ExecutionContext";
@@ -13,4 +13,3 @@ export * from "./parallelQueryExecutionContext";
 export * from "./orderByQueryExecutionContext";
 export * from "./pipelinedQueryExecutionContext";
 export * from "./orderByComparator";
-export * from "./documentProducer2";
