+++ sdk/cosmosdb/cosmos/src/request/FeedOptions.ts
@@ -130,8 +130,9 @@ export interface FeedOptions extends SharedOptions {
   allowUnboundedNonStreamingQueries?: boolean;
 
   /**
-   * Enable query control for the query.
-   * Would give empty results if the results is not ready to served.
+   * Controls query execution behavior.
+   * Default: false. If set to false, the query will retry until results are ready and `maxItemCount` is reached, which can take time for large partitions with relatively small data.
+   * If set to true, scans partitions up to `maxDegreeOfParallelism`, adds results to the buffer, and returns what is available. If results are not ready, it returns an empty response.
    */
   enableQueryControl?: boolean;
 }
